Prevents crash without Arc import
Before this modification, if one were to uncomment the smiley face for
the sun and tried to compile the code, the compiler would complain about
a lack of import for Arc. However, Arc was only used for the sun's face,
so it'd be an unused import if the face was commented out. To resolve
the problem, the import was directly called for the single case the Arc
was called.